refactor: modularize core and app config files
- moved config.py → app/config.py
- moved middleware.py → app/middleware.py
- moved sanitizer.py → core/security/sanitizer.py
- moved security_utils.py → core/security/security_utils.py
- moved security_parsers.py → core/security/security_parsers.py
- moved error_responses.py → core/responses/error_responses.py
- moved validators.py → core/validators/validators.py
- moved exceptions.py → core/exceptions/exceptions.py
- updated all import paths accordingly
