+++ backend/app.py
@@ -14,7 +14,7 @@ from flask_migrate import Migrate
 from flask_wtf.csrf import CSRFProtect, generate_csrf, CSRFError
 from sqlalchemy.exc import SQLAlchemyError
 from dotenv import load_dotenv
-from backend.spine_generator import create_spine
+from spine_generator import create_spine
 import os
 import requests
 
@@ -23,16 +23,16 @@ from datetime import datetime, timedelta, timezone
 import sys
 import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from sanitizer import sanitize_payload
+from core.security.sanitizer import sanitize_payload
 from reader_identity.routes import reader_identity_bp
 
 # Environment variables are now loaded centrally in backend/config.py
-from config import app_config, setup_logging, validate_required_env_vars
+from app.config import app_config, setup_logging, validate_required_env_vars
 from ai_service import generate_book_note, get_ai_recommendations, get_category_books, get_book_mood_tags_safe, generate_chat_response, llm_service
 from models import db, User, Book, ShelfItem, BookNote, ReadingGoal, ReadingStats, Collection, CollectionItem, PriceHistory, PriceAlert, Review, register_user, login_user
 from price_tracker import get_price_tracker
 from cache_service import cache_service
-from validators import (
+from core.validators.validators import (
     validate_request,
     validate_google_books_id,
     AnalyzeMoodRequest,
@@ -62,7 +62,7 @@ from validators import (
 from collections import defaultdict, deque
 from math import ceil
 from time import time
-from error_responses import (
+from core.responses.error_responses import (
     ErrorCodes, error_response, success_response,
     validation_error, missing_fields_error, invalid_json_error,
     auth_error, forbidden_error, unauthorized_access_error,
@@ -448,11 +448,11 @@ def handle_analyze_mood():
 @rate_limit('mood_tags')
 def handle_mood_tags():
     """Get mood tags for a book."""
-    from exceptions import (
+    from core.exceptions.exceptions import (
         LLMCircuitBreakerOpenError, AIServiceException, 
         ValidationException, InvalidInputError
     )
-    from error_responses import handle_exception
+    from core.responses.error_responses import handle_exception
     
     try:
         data = request.get_json()
@@ -483,11 +483,11 @@ def handle_mood_tags():
 @rate_limit('mood_search')
 def handle_mood_search():
     """Search for books based on mood/vibe with improved query parsing."""
-    from exceptions import (
+    from core.exceptions.exceptions import (
         LLMCircuitBreakerOpenError, AIServiceException,
         ValidationException, InvalidInputError
     )
-    from error_responses import handle_exception
+    from core.responses.error_responses import handle_exception
     
     try:
         data = request.get_json()
@@ -604,12 +604,12 @@ def handle_category_books():
 @rate_limit('generate_note')
 def handle_generate_note():
     """Generate AI-powered book recommendation with vibe support."""
-    from exceptions import (
+    from core.exceptions.exceptions import (
         LLMCircuitBreakerOpenError, AIServiceException,
         DatabaseQueryError, DatabaseIntegrityError,
         ValidationException, InvalidInputError
     )
-    from error_responses import handle_exception
+    from core.responses.error_responses import handle_exception
     
     try:
         data = request.get_json()
@@ -661,11 +661,11 @@ def handle_generate_note():
 @rate_limit('chat')
 def handle_chat():
     """Handle chat messages and generate bookseller responses."""
-    from exceptions import (
+    from core.exceptions.exceptions import (
         LLMCircuitBreakerOpenError, AIServiceException,
         ValidationException, InvalidInputError
     )
-    from error_responses import handle_exception
+    from core.responses.error_responses import handle_exception
     
     try:
         data = request.get_json()
@@ -756,8 +756,8 @@ def health_check():
 def add_to_library():
     """Add a book to the user's shelf."""
     from sqlalchemy.exc import IntegrityError
-    from exceptions import DatabaseQueryError, DatabaseIntegrityError, ValidationException
-    from error_responses import handle_exception
+    from core.exceptions.exceptions import DatabaseQueryError, DatabaseIntegrityError, ValidationException
+    from core.responses.error_responses import handle_exception
     
     try:
         data = request.get_json()
@@ -1196,8 +1196,8 @@ def register():
 @limiter.limit("5 per minute")
 def login():
     """Authenticate user and return JWT token."""
-    from exceptions import DatabaseQueryError, ValidationException
-    from error_responses import handle_exception
+    from core.exceptions.exceptions import DatabaseQueryError, ValidationException
+    from core.responses.error_responses import handle_exception
     
     try:
         data = request.get_json()
